Move implementation into lib file
